+++ example/index.ts
@@ -1,75 +1,31 @@
 import { t } from 'elysia'
 import { createMirror } from '../src/index'
+import { TypeCompiler } from '@sinclair/typebox/compiler'
 
 const shape = t.Object({
-	id: t.Number(),
-	name: t.Literal('SaltyAom'),
-	bio: t.String({
-		sanitize: true
-	}),
-	user: t.Object({
-		name: t.String(),
-		password: t.String()
-	}),
-	playing: t.Optional(t.String()),
-	games: t.Array(
-		t.Object({
-			name: t.String(),
-			hoursPlay: t.Number({ default: 0 }),
-			tags: t.Array(t.String())
-		})
-	),
-	metadata: t.Intersect([
-		t.Object({
-			alias: t.String()
-		}),
-		t.Object({
-			country: t.Nullable(t.String())
-		})
-	]),
-	social: t.Optional(
-		t.Object({
-			facebook: t.Optional(t.String()),
-			twitter: t.Optional(t.String()),
-			youtube: t.Optional(t.String())
-		})
-	)
+	a: t.Union([
+		t.Object({ a: t.String() }),
+		t.Object(
+			{ b: t.String() },
+			{
+				additionalProperties: true
+			}
+		),
+		t.String()
+	])
 })
 
 const value = {
-	id: 1,
-	name: 'SaltyAom',
-	bio: 'I like train\nhere',
-	user: {
-		name: 'SaltyAom',
-		password: '123456'
-	},
-	games: [
-		{
-			name: 'MiSide',
-			hoursPlay: 17,
-			tags: ['Psychological Horror', 'Cute', 'Dating Sim']
-		},
-		{
-			name: 'Strinova',
-			hoursPlay: 365,
-			tags: ['Free to Play', 'Anime', 'Third-Person Shooter']
-		},
-		{
-			name: "Tom Clancy's Rainbow Six Siege",
-			hoursPlay: 287,
-			tags: ['FPS', 'Multiplayer', 'Tactical']
-		}
-	],
-	metadata: {
-		alias: 'SaltyAom',
-		country: 'Thailand'
+	a: {
+		a: 'b',
+		c: 'a'
 	}
 } satisfies typeof shape.static
 
-const mirror = createMirror(shape)
+const mirror = createMirror(shape, {
+	// TypeCompiler
+})
 
 console.dir(mirror(value), {
 	depth: null
 })
-// console.log(stringify.toString())