+++ test/index.test.ts
@@ -1,11 +1,7 @@
 import { t } from 'elysia'
-import { type TAnySchema } from '@sinclair/typebox'
-import { createMirror } from '../src'
 
-import { describe, expect, it } from 'bun:test'
-
-const isEqual = (shape: TAnySchema, value: unknown, expected = value) =>
-	expect(createMirror(shape)(value)).toEqual(expected)
+import { describe, it, mock } from 'bun:test'
+import { isEqual } from './utils'
 
 describe('Core', () => {
 	it('handle string', () => {
@@ -116,44 +112,6 @@ describe('Core', () => {
 		isEqual(shape, value, expected)
 	})
 
-	it('handle nested object with optional', () => {
-		const shape = t.Object({
-			name: t.String(),
-			info: t.Object({
-				alias: t.String(),
-				optional1: t.Optional(t.String()),
-				optional2: t.Optional(t.String())
-			}),
-			optional1: t.Optional(t.String()),
-			optional2: t.Optional(t.String())
-		})
-
-		const value = {
-			name: 'salt',
-			additional: 'b',
-			info: {
-				alias: 'salty',
-				// @ts-expect-error
-				additional: 'b',
-				optional2: 'ok'
-			},
-			optional1: 'ok',
-			// @ts-expect-error
-			additional: 'b'
-		} satisfies typeof shape.static
-
-		const expected = {
-			name: 'salt',
-			info: {
-				alias: 'salty',
-				optional2: 'ok'
-			},
-			optional1: 'ok'
-		} satisfies typeof shape.static
-
-		isEqual(shape, value, expected)
-	})
-
 	it('handle array object', () => {
 		const shape = t.Array(
 			t.Object({
@@ -230,92 +188,34 @@ describe('Core', () => {
 		isEqual(shape, value, expected)
 	})
 
-	it('handle array object with optional', () => {
-		const shape = t.Array(
+	it('handle intersect object', () => {
+		const shape = t.Intersect([
 			t.Object({
-				name: t.String(),
-				optional1: t.Optional(t.String()),
-				optional2: t.Optional(t.String())
-			})
-		)
-
-		const value = [
-			{
-				name: 'salt',
-				optional1: 'ok',
-				// @ts-expect-error
-				additional: 'b'
-			},
-			{
-				name: 'chiffon'
-			}
-		] satisfies typeof shape.static
-
-		const expected = [
-			{
-				name: 'salt',
-				optional1: 'ok'
-			},
-			{
-				name: 'chiffon'
-			}
-		] satisfies typeof shape.static
-
-		isEqual(shape, value, expected)
-	})
-
-	it('handle array nested object with optional', () => {
-		const shape = t.Array(
+				name: t.String()
+			}),
 			t.Object({
-				name: t.String(),
 				info: t.Object({
-					alias: t.String(),
-					optional1: t.Optional(t.String()),
-					optional2: t.Optional(t.String())
-				}),
-				optional1: t.Optional(t.String()),
-				optional2: t.Optional(t.String())
+					alias: t.String()
+				})
 			})
-		)
+		])
 
-		const value = [
-			{
-				name: 'salt',
-				additional: 'b',
-				info: {
-					alias: 'salty',
-					// @ts-expect-error
-					additional: 'b',
-					optional2: 'ok'
-				},
-				optional1: 'ok',
+		const value = {
+			name: 'salt',
+			additional: 'b',
+			info: {
+				alias: 'salty',
 				// @ts-expect-error
 				additional: 'b'
-			},
-			{
-				name: 'chiffon',
-				info: {
-					alias: 'chiffon'
-				}
 			}
-		] satisfies typeof shape.static
+		} satisfies typeof shape.static
 
-		const expected = [
-			{
-				name: 'salt',
-				info: {
-					alias: 'salty',
-					optional2: 'ok'
-				},
-				optional1: 'ok'
-			},
-			{
-				name: 'chiffon',
-				info: {
-					alias: 'chiffon'
-				}
+		const expected = {
+			name: 'salt',
+			info: {
+				alias: 'salty'
 			}
-		] satisfies typeof shape.static
+		} satisfies typeof shape.static
 
 		isEqual(shape, value, expected)
 	})